+++ work/app/_parts/_footer.php
@@ -1,0 +1,2 @@
+</body>
+</html>
