+++ work/app/_parts/_header.php
@@ -1,0 +1,7 @@
+<!DOCTYPE html>
+<html lang="ja">
+<head>
+  <meta charset="utf-8">
+  <title>PHP Practice</title>
+</head>
+<body>
