+++ work/app/functions.php
@@ -1,0 +1,6 @@
+<?php
+
+function h($str)
+{
+  return htmlspecialchars($str, ENT_QUOTES, 'UTF-8');
+}
